Other Entropies

Supplementary functions for various tasks related to EntropyHub and signal processing.

EntropyHub._ExampleData.ExampleData — Function

Data = ExampleData(SigName::String)

Imports sample data time series with specific properties that are commonly used as benchmarks for assessing the performance of various entropy methods. The datasets returned by ExampleData() are used in the examples provided in documentation on www.EntropyHub.xyz and elsewhere. ***Note*** ExampleData() requires an internet connection to download and import the required datasets!

Datais the sample dataset imported corresponding to the string input SigName which can be one of the following string:

Arguments:

SigName -

        `uniform`          - uniformly distributed random number sequence in range [0 1], N = 5000
        `randintegers`     - randomly distributed integer sequence in range [1 8], N = 4096
        `gaussian`         - normally distributed number sequence [mean: 0, SD: 1], N = 5000
        `henon`            - X and Y components of the Henon attractor [alpha: 1.4, beta: .3, Xo = 0, Yo = 0], N = 4500
        `lorenz`           - X, Y, and Z components of the Lorenz attractor [sigma: 10, beta: 8/3, rho: 28, Xo = 10, Yo = 20, Zo = 10], N = 5917
        `chirp`            - chirp signal (f0 = .01, t1 = 4000, f1 = .025), N = 5000
        `uniform2`         - two uniformly distributed random number sequences in range [0,1], N = 4096
        `gaussian2`        - two normally distributed number sequences [mean: 0, SD: 1], N = 3000
        `randintegers2`    - two uniformly distributed pseudorandom integer sequences in range [1 8], N = 3000
        `uniform_Mat`      - matrix of uniformly distributed random numbers in range [0 1], N = 50 x 50
        `gaussian_Mat`     - matrix of normally distributed numbers [mean: 0, SD: 1], N = 60 x 120
        `randintegers_Mat` - matrix of randomly distributed integers in range [1 8], N = 88 x 88
        `mandelbrot_Mat`   - matrix representing a Mandelbrot fractal image with values in range [0 255], N = 92 x 115
        `entropyhub_Mat`   - matrix representing the EntropyHub logo with values in range [0 255], N = 127 x 95
                     
For further info on these graining procedures see the `EntropyHub guide <https://github.com/MattWillFlood/EntropyHub/blob/main/EntropyHub%20Guide.pdf>`_.

EntropyHub._WindowData.WindowData — Function

WinData, Log = WindowData(Data)

Windows the sequence(s) given in Data into a collection of subsequnces of floor(N/5) elements with no overlap, excluding any remainder elements that do not fill the final window. If Data is a univariate sequence (vector), Windata is a vector of 5 vectors. If Data is a set of multivariate sequences (NxM matrix), each of M columns is treated as a sequence with N elements and WinData is a vector of 5 matrices of size [(floor*N,5), M]. The Log dictionary contains information about the windowing process, including: DataType - The type of data sequence passed as Data

DataLength - The number of sequence elements in Data

WindowLength - The number of elements in each window of WinData

WindowOverlap - The number of overlapping elements between windows

TotalWindows - The number of windows extracted from Data

Mode - Decision to include or exclude any remaining sequence elements (< WinLen) that do not fill the window.

WinData, Log = WindowData(Data::AbstractArray{T} where T<:Real, WinLen::Union{Nothing,Int}=nothing, Overlap::Int=0, Mode::String="exclude")

Windows the sequence(s) given in Data into a collection of subsequnces using the specified keyword arguments:

Arguments:

WinLen - Number of elements in each window, a positive integer (>10)

Overlap - Number of overlapping elements between windows, a positive integer (< WinLen)

Mode - Decision to include or exclude any remaining sequence elements (< WinLen) that do not fill the window, a string - either "include" or "exclude" (default).

See also ExampleData

EntropyHub.MSobject